% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Annotation-class}
\alias{Annotation-class}
\title{Annotation Class}
\description{
Class for storing cell type annotations and their numeric encodings
}
\section{Slots}{

\describe{
\item{\code{ref_celltype}}{Vector of cell type labels}

\item{\code{ref_celltype.code}}{Numeric vector of encoded cell types}
}}

