% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_a_matrix}
\alias{calculate_a_matrix}
\title{Calculate A Matrix}
\usage{
calculate_a_matrix(counts)
}
\arguments{
\item{counts}{Cell type counts}
}
\value{
A matrix
}
\description{
Calculate A Matrix
}
\keyword{internal}
