% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_accuracy}
\alias{calculate_accuracy}
\title{Calculate Accuracy}
\usage{
calculate_accuracy(train_object, h_project, seed = 1, return_pred = FALSE)
}
\arguments{
\item{train_object}{Training object}

\item{h_project}{Projected H matrix}

\item{seed}{Random seed}

\item{return_pred}{Return prediction details}
}
\value{
Accuracy or list with accuracy and predictions
}
\description{
Calculate Accuracy
}
\keyword{internal}
