% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_b_matrix}
\alias{calculate_b_matrix}
\title{Calculate B Matrix}
\usage{
calculate_b_matrix(counts)
}
\arguments{
\item{counts}{Cell type counts}
}
\value{
B matrix
}
\description{
Calculate B Matrix
}
\keyword{internal}
