% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_bp_posneg}
\alias{calculate_bp_posneg}
\title{Calculate BP Positive/Negative Components}
\usage{
calculate_bp_posneg(B, P)
}
\arguments{
\item{B}{B matrix}

\item{P}{P matrix}
}
\value{
List of positive and negative components
}
\description{
Calculate BP Positive/Negative Components
}
\keyword{internal}
