% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_const_for_h}
\alias{calculate_const_for_h}
\title{Calculate Constants for H}
\usage{
calculate_const_for_h(object, beta = NULL)
}
\arguments{
\item{object}{CSFNMF object}

\item{beta}{Optional beta matrix}
}
\value{
List of constants for H updates
}
\description{
Calculate Constants for H
}
\keyword{internal}
