% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainModelCSFNMF.R
\name{calculate_correlation_matrix}
\alias{calculate_correlation_matrix}
\title{Calculate Correlation Matrix}
\usage{
calculate_correlation_matrix(true_labels, pred_labels)
}
\arguments{
\item{true_labels}{True cell type labels}

\item{pred_labels}{Predicted cell type labels}
}
\value{
Correlation matrix
}
\description{
Calculate Correlation Matrix
}
\keyword{internal}
