% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_help_matrices}
\alias{calculate_help_matrices}
\title{Calculate Help Matrices}
\usage{
calculate_help_matrices(object)
}
\arguments{
\item{object}{CSFNMF object}
}
\value{
Updated object with calculated matrices
}
\description{
Calculates helper matrices for NMF optimization
}
\keyword{internal}
