% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{calculate_loss}
\alias{calculate_loss}
\title{Calculate Loss Function}
\usage{
calculate_loss(object, divided = FALSE)
}
\arguments{
\item{object}{CSFNMF object}

\item{divided}{Logical: return components separately}
}
\value{
Loss value or list of components
}
\description{
Calculate Loss Function
}
\keyword{internal}
