% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_n_matrix}
\alias{calculate_n_matrix}
\title{Calculate N Matrix}
\usage{
calculate_n_matrix(celltype)
}
\arguments{
\item{celltype}{Cell type vector}
}
\value{
N matrix
}
\description{
Calculate N Matrix
}
\keyword{internal}
