% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_nmi}
\alias{calculate_nmi}
\title{Calculate Normalized Mutual Information}
\usage{
calculate_nmi(true_labels, pred_labels)
}
\arguments{
\item{true_labels}{Vector of true labels}

\item{pred_labels}{Vector of predicted labels}
}
\value{
NMI score between 0 and 1
}
\description{
Calculate Normalized Mutual Information
}
\keyword{internal}
