% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_p_matrix}
\alias{calculate_p_matrix}
\title{Calculate P Matrix}
\usage{
calculate_p_matrix(num_clusters)
}
\arguments{
\item{num_clusters}{Number of clusters}
}
\value{
P matrix
}
\description{
Calculate P Matrix
}
\keyword{internal}
