% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_performance}
\alias{calculate_performance}
\title{Calculate Performance Metrics using Clustering}
\usage{
calculate_performance(
  train_object,
  h_project,
  seed = 1,
  return_pred = FALSE,
  method = "snn"
)
}
\arguments{
\item{train_object}{Training object}

\item{h_project}{Projected H matrix}

\item{seed}{Random seed}

\item{return_pred}{Return prediction details}

\item{method}{Clustering method to use}
}
\value{
List containing NMI score and optionally cluster assignments
}
\description{
Calculate Performance Metrics using Clustering
}
\keyword{internal}
