% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectRank.R
\name{calculate_significant_components}
\alias{calculate_significant_components}
\title{Optimized Component Significance Calculation}
\usage{
calculate_significant_components(vectors, k, n.genes, max_p_value)
}
\description{
Optimized Component Significance Calculation
}
\keyword{internal}
