% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeWH.R
\name{calculate_similarity_matrix}
\alias{calculate_similarity_matrix}
\title{Calculate Similarity Matrix}
\usage{
calculate_similarity_matrix(data_matrix, centroids, cell_clusters)
}
\arguments{
\item{data_matrix}{Data matrix}

\item{centroids}{Centroid matrix}

\item{cell_clusters}{Cell cluster assignments}
}
\value{
Similarity matrix
}
\description{
Calculate Similarity Matrix
}
\keyword{internal}
