% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\name{CITEseqDataset}
\alias{CITEseqDataset}
\title{Load CITE-seq Dataset}
\usage{
CITEseqDataset()
}
\value{
A list containing:
\item{data}{RNA expression matrix}
\item{ADT}{Antibody-derived tag expression matrix}
\item{ADT scale}{Scaled antibody-derived tag data}
}
\description{
Loads and processes CITE-seq PBMC dataset with protein expression data
}
