% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{cluster_nmf}
\alias{cluster_nmf}
\title{Perform clustering on NMF factors}
\usage{
cluster_nmf(h_matrix, n_clusters, method = "kmeans", resolution = 0.8)
}
\arguments{
\item{h_matrix}{H matrix from NMF}

\item{n_clusters}{Number of clusters to find}

\item{method}{Clustering method ('kmeans', 'hclust', or 'leiden')}
}
\value{
Vector of cluster assignments
}
\description{
Perform clustering on NMF factors
}
\keyword{internal}
