% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{cm_annotation}
\alias{cm_annotation}
\title{Access cell annotations from a CellMentor object}
\usage{
cm_annotation(x)
}
\arguments{
\item{x}{A CellMentor object.}
}
\value{
A data frame containing cell annotations.
}
\description{
Retrieve cell type or metadata annotations.
}
\examples{
data(obj_toy, package = "CellMentor")
cm_annotation(obj_toy)
}
