% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{cm_rank}
\alias{cm_rank}
\title{Access the rank of a CellMentor object}
\usage{
cm_rank(x)
}
\arguments{
\item{x}{A CellMentor object.}
}
\value{
A numeric value representing the selected rank.
}
\description{
Retrieve the factorization rank used during CSFNMF training.
}
\examples{
\dontrun{
# Access the rank of the model
cm_rank(cs_obj)
}
}
