% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_cache}
\alias{create_cache}
\title{Cache Manager}
\usage{
create_cache(max_size = 1000)
}
\arguments{
\item{max_size}{Maximum cache size in MB}
}
\value{
Cache environment
}
\description{
Manages temporary results caching
}
\keyword{internal}
