% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_progress}
\alias{create_progress}
\title{Create Progress Manager}
\usage{
create_progress(total, title = "Progress")
}
\arguments{
\item{total}{Total steps}

\item{title}{Progress bar title}
}
\value{
Progress manager object
}
\description{
Create Progress Manager
}
\keyword{internal}
