% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_reporter}
\alias{create_reporter}
\title{Progress Reporter Constructor}
\usage{
create_reporter(verbose = TRUE, with_timing = TRUE)
}
\arguments{
\item{verbose}{Logical: show messages}

\item{with_timing}{Logical: include timestamps}
}
\value{
Function for reporting progress
}
\description{
Creates a progress reporting function with optional timing
}
\keyword{internal}
