% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_functions.R
\name{create_train_object}
\alias{create_train_object}
\title{Create Training Object for CSFNMF}
\usage{
create_train_object(ref_matrix, ref_celltype, data_matrix, data_celltype)
}
\arguments{
\item{ref_matrix}{Reference matrix}

\item{ref_celltype}{Reference cell types}

\item{data_matrix}{Data matrix}

\item{data_celltype}{Data cell types}
}
\value{
TrainCSFNMF object
}
\description{
Create Training Object for CSFNMF
}
\keyword{internal}
