% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{csfnmf-class}
\alias{csfnmf-class}
\title{CSFNMF Class}
\description{
Main class for Constrained Supervised Factorization NMF analysis
}
\section{Slots}{

\describe{
\item{\code{count.matrices}}{Current working matrices}

\item{\code{origin.matrices}}{Original unprocessed matrices}

\item{\code{norm.matrices}}{Normalized matrices}

\item{\code{scale.matrices}}{Scaled matrices}

\item{\code{f.select.matrices}}{Feature-selected matrices}

\item{\code{annotation}}{Data frame of annotations}

\item{\code{train_object}}{Trained model parameters}

\item{\code{H}}{Factor loading matrix for cells}

\item{\code{W}}{Factor loading matrix for genes}
}}

