% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{CSFnmfSingleR}
\alias{CSFnmfSingleR}
\title{SingleR Prediction on CSFNMF}
\usage{
CSFnmfSingleR(train_object, H_data = NULL, de.n = 50)
}
\arguments{
\item{train_object}{Training object}

\item{H_data}{Optional H matrix}

\item{de.n}{Number of genes for DE}
}
\value{
SingleR predictions
}
\description{
SingleR Prediction on CSFNMF
}
\keyword{internal}
