% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{data_matrix}
\alias{data_matrix}
\title{Access the query (data) matrix from a RefDataList object}
\usage{
data_matrix(x)
}
\arguments{
\item{x}{A RefDataList object.}
}
\value{
A sparse \code{Matrix::Matrix} object representing query data.
}
\description{
Retrieve the single-cell expression matrix for the query dataset.
}
\examples{
data(obj_toy, package = "CellMentor")
data_matrix(matrices(obj_toy))
}
