% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectRank.R
\name{debug_matrix}
\alias{debug_matrix}
\title{Print Matrix Info}
\usage{
debug_matrix(mat, name = "Matrix")
}
\arguments{
\item{mat}{Matrix to examine}

\item{name}{Name for output}
}
\description{
Print Matrix Info
}
\keyword{internal}
