% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_functions.R
\name{divide_reference_data}
\alias{divide_reference_data}
\title{Divide Reference Data for Training}
\usage{
divide_reference_data(object, seed = 1)
}
\arguments{
\item{object}{CSFNMF object}

\item{seed}{Random seed}
}
\value{
Training object
}
\description{
Divide Reference Data for Training
}
\keyword{internal}
