% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{encode_celltypes}
\alias{encode_celltypes}
\title{Encode Cell Types}
\usage{
encode_celltypes(celltype_list)
}
\arguments{
\item{celltype_list}{Vector of cell type labels}
}
\value{
Factor with numeric levels
}
\description{
Converts cell type labels to numeric codes efficiently
}
\keyword{internal}
