% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_csfnmf.R
\name{ensure_dgCMatrix}
\alias{ensure_dgCMatrix}
\title{Ensure Matrix Format is dgCMatrix}
\usage{
ensure_dgCMatrix(mat)
}
\arguments{
\item{mat}{Input matrix}
}
\value{
Matrix in dgCMatrix format
}
\description{
Ensure Matrix Format is dgCMatrix
}
\keyword{internal}
