% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectRank.R
\name{ensure_matrix}
\alias{ensure_matrix}
\title{Ensure Matrix Format}
\usage{
ensure_matrix(mat)
}
\arguments{
\item{mat}{Input matrix}
}
\value{
Properly formatted sparse matrix
}
\description{
Ensure Matrix Format
}
\keyword{internal}
