% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{find_common_genes}
\alias{find_common_genes}
\title{Find Common Genes}
\usage{
find_common_genes(ref_matrix, data_matrix)
}
\arguments{
\item{ref_matrix}{Reference expression matrix}

\item{data_matrix}{Query expression matrix}
}
\value{
List of indices and shared genes
}
\description{
Identifies common genes between reference and query datasets
}
\keyword{internal}
