% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{H}
\alias{H}
\title{Access the H matrix}
\usage{
H(x)
}
\arguments{
\item{x}{A CellMentor object.}
}
\value{
A numeric matrix representing the H (cell embeddings) matrix.
}
\description{
Retrieve the H matrix (cell embeddings) from a CellMentor object.
}
\examples{
data(obj_toy, package = "CellMentor")
H(obj_toy)
}
