% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{helpmat-class}
\alias{helpmat-class}
\title{Helper Matrices Class}
\description{
Class for storing intermediate matrices used in NMF computation
}
\section{Slots}{

\describe{
\item{\code{A}}{First matrix}

\item{\code{B}}{Second matrix}

\item{\code{P}}{P matrix}

\item{\code{M}}{M matrix}

\item{\code{N}}{N matrix}

\item{\code{BP}}{BP matrix}

\item{\code{BP_posneg}}{List of positive/negative BP values}

\item{\code{Hconst}}{List of H constants}
}}

