% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeWH.R
\name{initialize_cells_cluster}
\alias{initialize_cells_cluster}
\title{Initialize Matrices Using Cell Clustering}
\usage{
initialize_cells_cluster(data_matrix, k, labels)
}
\arguments{
\item{data_matrix}{Data matrix}

\item{k}{Rank}

\item{labels}{Cell type labels}
}
\value{
List containing W and H matrices
}
\description{
Initialize Matrices Using Cell Clustering
}
\keyword{internal}
