% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeWH.R
\name{initialize_genes_cluster}
\alias{initialize_genes_cluster}
\title{Initialize Matrices Using Gene Clustering}
\usage{
initialize_genes_cluster(data_matrix, k, labels)
}
\arguments{
\item{data_matrix}{Data matrix}

\item{k}{Rank}

\item{labels}{Cell type labels}
}
\value{
List containing W and H matrices
}
\description{
Initialize Matrices Using Gene Clustering
}
\keyword{internal}
