% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeWH.R
\name{initialize_nndsvd}
\alias{initialize_nndsvd}
\title{Initialize Using NNDSVD Method}
\usage{
initialize_nndsvd(data_matrix, k)
}
\arguments{
\item{data_matrix}{Input matrix}

\item{k}{Rank}
}
\value{
List containing W and H matrices
}
\description{
Initialize Using NNDSVD Method
}
\keyword{internal}
