% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeWH.R
\name{initialize_regulated}
\alias{initialize_regulated}
\title{Initialize Matrices Using Up/Down Regulated Method}
\usage{
initialize_regulated(data_matrix, k, labelcode)
}
\arguments{
\item{data_matrix}{Data matrix}

\item{k}{Rank}

\item{labelcode}{Cell type codes}
}
\value{
List containing W and H matrices
}
\description{
Initialize Matrices Using Up/Down Regulated Method
}
\keyword{internal}
