% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeWH.R
\name{initialize_uniform}
\alias{initialize_uniform}
\title{Initialize Matrices Using Uniform Distribution}
\usage{
initialize_uniform(data_matrix, k)
}
\arguments{
\item{data_matrix}{Data matrix}

\item{k}{Rank}
}
\value{
List containing W and H matrices
}
\description{
Initialize Matrices Using Uniform Distribution
}
\keyword{internal}
