% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeWH.R
\name{initialize_wh}
\alias{initialize_wh}
\title{Initialize W and H Matrices}
\usage{
initialize_wh(object, method = "NNDSVD", seed = 1)
}
\arguments{
\item{object}{CSFNMF object}

\item{method}{Initialization method ("uniform", "regulated", "NNDSVD", "skmeanGenes", "skmeanCells")}

\item{seed}{Random seed}
}
\value{
List containing W and H matrices
}
\description{
Initializes W and H matrices using various methods
}
