% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading.R
\name{load_pbmcsca_data}
\alias{load_pbmcsca_data}
\title{Load PBMC Dataset from SeuratData}
\usage{
load_pbmcsca_data()
}
\value{
A list containing processed data and cell types
}
\description{
Load PBMC Dataset from SeuratData
}
