% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{matrices}
\alias{matrices}
\title{Access the RefDataList from a CSFNMF object}
\usage{
matrices(x)
}
\arguments{
\item{x}{A \code{csfnmf} or \code{traincsfnmf} object.}
}
\value{
A RefDataList object containing the reference and query matrices.
}
\description{
Retrieve the RefDataList structure containing both reference and
query matrices.
}
\examples{
data(obj_toy, package = "CellMentor")
matrices(obj_toy)
}
