% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectRank.R
\name{NonSquareSinkhornKnopp}
\alias{NonSquareSinkhornKnopp}
\title{Corrected Sinkhorn-Knopp Matrix Scaling}
\usage{
NonSquareSinkhornKnopp(
  A,
  sum_row = matrix(ncol(A), nrow = nrow(A), ncol = 1),
  sum_col = matrix(nrow(A), nrow = ncol(A), ncol = 1),
  niter = 100,
  tol = 1e-10,
  alpha = NULL,
  beta = NULL
)
}
\arguments{
\item{A}{Input matrix}

\item{sum_row}{Row sums}

\item{sum_col}{Column sums}

\item{niter}{Maximum iterations}

\item{tol}{Convergence tolerance}

\item{alpha}{Alpha parameter}

\item{beta}{Beta parameter}
}
\value{
List of scaling vectors
}
\description{
Corrected Sinkhorn-Knopp Matrix Scaling
}
\keyword{internal}
