% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{normalize_matrix}
\alias{normalize_matrix}
\title{Normalize Matrix}
\usage{
normalize_matrix(matrix, scale.factor = 10000)
}
\arguments{
\item{matrix}{Input matrix}

\item{scale.factor}{Scale factor for normalization}
}
\value{
Normalized matrix
}
\description{
Performs log normalization on expression matrix
}
\keyword{internal}
