% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R, R/data-toy.R
\docType{data}
\name{obj_toy}
\alias{obj_toy}
\title{Tiny toy CSFNMF object for examples}
\format{
A \code{csfnmf} object with:
\describe{
\item{\code{matrices@ref}}{dgCMatrix (genes x cells), 30 x 60}
\item{\code{matrices@data}}{dgCMatrix (genes x cells), 30 x 40}
\item{\code{annotation}}{data.frame with celltype}
}

An object of class \code{csfnmf} built on the toy matrices.
}
\usage{
data(obj_toy)

obj_toy
}
\value{
toy data in proper format
}
\description{
A small object for fast runnable examples and tests.
}
\details{
Only provided to make accessor examples immediate. For real analyses,
construct objects from your own data.
}
\examples{
data(obj_toy)
stopifnot(inherits(obj_toy, "csfnmf"))
\donttest{
data(obj_toy, package = "CellMentor")
inherits(obj_toy, "csfnmf")
}
}
\keyword{datasets}
