% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectRank.R
\name{OptimalBeta}
\alias{OptimalBeta}
\title{Optimized Optimal Beta Calculation}
\usage{
OptimalBeta(data, beta)
}
\arguments{
\item{data}{Input matrix}

\item{beta}{Beta value}
}
\value{
Difference statistic
}
\description{
Optimized Optimal Beta Calculation
}
\keyword{internal}
