% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parallel_process}
\alias{parallel_process}
\title{Process in Parallel}
\usage{
parallel_process(X, FUN, num_cores = 1, chunk_size = NULL)
}
\arguments{
\item{X}{Data to process}

\item{FUN}{Function to apply}

\item{num_cores}{Number of cores}

\item{chunk_size}{Size of chunks}
}
\value{
Results list
}
\description{
Process in Parallel
}
\keyword{internal}
