% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridSearch.R
\name{plot_parameter_search}
\alias{plot_parameter_search}
\title{Plot Parameter Search Results}
\usage{
plot_parameter_search(results)
}
\arguments{
\item{results}{Results from select_optimal_parameters}
}
\value{
ggplot object with visualization of parameter search results
}
\description{
Plot Parameter Search Results
}
