% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subclustering.R
\name{predict_subclusters}
\alias{predict_subclusters}
\title{Predict Sub-clusters for Query Data}
\usage{
predict_subclusters(object, verbose = TRUE)
}
\arguments{
\item{object}{CSFNMF object}

\item{verbose}{Show progress}
}
\value{
Vector of predicted sub-clusters
}
\description{
Predict Sub-clusters for Query Data
}
\keyword{internal}
