% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_functions.R
\name{project_data}
\alias{project_data}
\title{Project Data onto Basis}
\usage{
project_data(W, X, seed = 1, num_cores = 1, chunk_size = 1000, verbose = TRUE)
}
\arguments{
\item{W}{Basis matrix}

\item{X}{Data matrix}

\item{seed}{Random seed}

\item{num_cores}{Number of cores for parallel processing}

\item{chunk_size}{Size of chunks for processing}

\item{verbose}{Show progress bar}
}
\value{
Projection matrix
}
\description{
Project Data onto Basis
}
\keyword{internal}
