% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{ref_matrix}
\alias{ref_matrix}
\title{Access the reference matrix from a RefDataList object}
\usage{
ref_matrix(x)
}
\arguments{
\item{x}{A RefDataList object.}
}
\value{
A sparse \code{Matrix::Matrix} object representing reference data.
}
\description{
Retrieve the single-cell expression matrix for the reference dataset.
}
\examples{
data(obj_toy, package = "CellMentor")
ref_matrix(matrices(obj_toy))
}
