% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R, R/classes.R
\docType{class}
\name{RefDataList-class}
\alias{RefDataList-class}
\title{RefDataList Class}
\description{
Class for storing paired reference and query expression matrices. The reference matrix contains
gene expression data from labeled cells, while the data matrix contains expression data from
unlabeled cells that need to be classified.

Class for storing paired reference and query expression matrices
}
\section{Slots}{

\describe{
\item{\code{ref}}{Reference matrix (genes × cells) containing expression data from labeled cells}

\item{\code{data}}{Query matrix (genes × cells) containing expression data from unlabeled cells}

\item{\code{ref}}{Reference matrix}

\item{\code{data}}{Query matrix}
}}

