% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{reorder_data}
\alias{reorder_data}
\title{Reorder Reference Data}
\usage{
reorder_data(object, reorder_by)
}
\arguments{
\item{object}{CSFNMF object}

\item{reorder_by}{Column to order by}
}
\value{
Updated object
}
\description{
Reorders reference data by cell types efficiently
}
\keyword{internal}
