% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{scale_by_genes}
\alias{scale_by_genes}
\title{Scale Matrix by Genes}
\usage{
scale_by_genes(matrix, eps = 1e-10)
}
\arguments{
\item{matrix}{Input matrix}

\item{eps}{Small value to prevent division by zero}
}
\value{
Scaled matrix
}
\description{
Scales matrix by genes using RMSD
}
\keyword{internal}
