% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{scale_rmsd}
\alias{scale_rmsd}
\title{Scale Matrix by RMSD}
\usage{
scale_rmsd(matrix, dimension = c("row", "column"), eps = 1e-10)
}
\arguments{
\item{matrix}{Input matrix}

\item{dimension}{"row" or "column"}

\item{eps}{Small value to prevent division by zero}
}
\value{
Scaled matrix
}
\description{
Scales matrix by root mean square deviation
}
\keyword{internal}
