% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{select_variable_genes}
\alias{select_variable_genes}
\title{Select Variable Genes}
\usage{
select_variable_genes(object, gene_list = NULL, num_cores = 1)
}
\arguments{
\item{object}{CSFNMF object}

\item{gene_list}{Additional genes to include}

\item{num_cores}{Number of cores for parallel processing}
}
\value{
List of updated matrices
}
\description{
Identifies highly variable genes using SingleR
}
\keyword{internal}
