% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectRank.R
\name{SelectAlphaForConstBeta}
\alias{SelectAlphaForConstBeta}
\title{Select Alpha for Constant Beta}
\usage{
SelectAlphaForConstBeta(Y, beta)
}
\arguments{
\item{Y}{Input matrix}

\item{beta}{Beta value}
}
\value{
List with alpha and scale matrix
}
\description{
Select Alpha for Constant Beta
}
\keyword{internal}
