% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{standard_beta}
\alias{standard_beta}
\title{Calculate Standard Beta Matrix}
\usage{
standard_beta(object, const = 1)
}
\arguments{
\item{object}{CSFNMF object}

\item{const}{Constant value for diagonal (default = 1)}
}
\value{
Beta matrix
}
\description{
Creates standardized beta matrix for NMF calculations
}
\keyword{internal}
