% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_sparse}
\alias{to_sparse}
\title{Safe Matrix Conversion}
\usage{
to_sparse(matrix, verbose = FALSE)
}
\arguments{
\item{matrix}{Input matrix}

\item{verbose}{Logical: show warnings}
}
\value{
Sparse matrix
}
\description{
Safely converts matrix to sparse format with error handling
}
\keyword{internal}
