% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainModelCSFNMF.R
\name{TrainModelCSFNMF}
\alias{TrainModelCSFNMF}
\title{Train CSFNMF Model}
\usage{
TrainModelCSFNMF(
  object,
  pra_to_update = c("alpha", "beta"),
  re_init = FALSE,
  num_update = 10,
  constSeed = 1,
  verbose = TRUE,
  theta = 1e-08,
  num_cores = 1
)
}
\arguments{
\item{object}{CSFNMF object}

\item{pra_to_update}{Parameters to update ("alpha", "beta")}

\item{re_init}{Reinitialize matrices}

\item{num_update}{Number of updates}

\item{constSeed}{Random seed}

\item{verbose}{Show progress}

\item{theta}{Convergence threshold}

\item{num_cores}{Number of cores for parallel processing}
}
\value{
Updated CSFNMF object
}
\description{
Train CSFNMF Model
}
