% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainModelCSFNMF.R
\name{update_alpha}
\alias{update_alpha}
\title{Update Alpha Parameter}
\usage{
update_alpha(train_object, correlation)
}
\arguments{
\item{train_object}{Training object}

\item{correlation}{Correlation matrix}
}
\value{
Updated alpha matrix
}
\description{
Update Alpha Parameter
}
\keyword{internal}
