% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainModelCSFNMF.R
\name{update_beta}
\alias{update_beta}
\title{Update Beta Parameter}
\usage{
update_beta(train_object, correlation)
}
\arguments{
\item{train_object}{Training object}

\item{correlation}{Correlation matrix}
}
\value{
Updated beta matrix
}
\description{
Update Beta Parameter
}
\keyword{internal}
