% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{update_h}
\alias{update_h}
\title{Update H Matrix}
\usage{
update_h(object, num_cores = 1)
}
\arguments{
\item{object}{CSFNMF object}

\item{num_cores}{Number of cores for parallel processing}
}
\value{
Updated H matrix
}
\description{
Update H Matrix
}
\keyword{internal}
