% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_progress}
\alias{update_progress}
\title{Update Progress}
\usage{
update_progress(progress, message = "", show_memory = TRUE)
}
\arguments{
\item{progress}{Progress manager}

\item{message}{Progress message}

\item{show_memory}{Show memory usage}
}
\description{
Update Progress
}
\keyword{internal}
