% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{update_training_performance}
\alias{update_training_performance}
\title{Update Training Object Performance with Clustering Results}
\usage{
update_training_performance(
  train_object,
  h_project,
  num_cores = 1,
  method = "kmeans"
)
}
\arguments{
\item{train_object}{Training object}

\item{h_project}{Projected H matrix}

\item{num_cores}{Number of cores for parallel processing}

\item{method}{Clustering method to use}
}
\value{
Updated training object with performance metrics
}
\description{
Update Training Object Performance with Clustering Results
}
\keyword{internal}
