% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{update_w}
\alias{update_w}
\title{Update W Matrix}
\usage{
update_w(object)
}
\arguments{
\item{object}{CSFNMF object}
}
\value{
Updated W matrix
}
\description{
Update W Matrix
}
\keyword{internal}
