% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_functions.R
\name{update_wh}
\alias{update_wh}
\title{Update W and H Matrices}
\usage{
update_wh(object, theta, verbose = TRUE, num_cores = 1)
}
\arguments{
\item{object}{CSFNMF object}

\item{theta}{Convergence threshold}

\item{verbose}{Show progress}

\item{num_cores}{Number of cores for parallel processing}
}
\value{
Updated object
}
\description{
Update W and H Matrices
}
\keyword{internal}
