% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_object.R
\name{validate_final_object}
\alias{validate_final_object}
\title{Validate Final CSFNMF Object}
\usage{
validate_final_object(object)
}
\arguments{
\item{object}{CSFNMF object to validate}
}
\value{
Logical indicating if object is valid
}
\description{
Validate Final CSFNMF Object
}
\keyword{internal}
