% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Input Validation}
\usage{
validate_inputs(ref_matrix, ref_celltype, data_matrix, params)
}
\arguments{
\item{ref_matrix}{Reference matrix}

\item{ref_celltype}{Reference cell types}

\item{data_matrix}{Query matrix}

\item{params}{List of additional parameters}
}
\value{
TRUE if valid, stops with error otherwise
}
\description{
Validates input parameters for CreateCSFNMFobject
}
\keyword{internal}
