% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{W}
\alias{W}
\title{Access the W matrix}
\usage{
W(x)
}
\arguments{
\item{x}{A CellMentor object (e.g., \code{traincsfnmf} or \code{csfnmf}).}
}
\value{
A numeric matrix representing the W (gene loadings) matrix.
}
\description{
Retrieve the W matrix (gene loadings) from a CellMentor object.
}
\examples{
data(obj_toy, package = "CellMentor")
W(obj_toy)
}
