% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_memory_check}
\alias{with_memory_check}
\title{Memory-Safe Operation}
\usage{
with_memory_check(expr, threshold = 10000)
}
\arguments{
\item{expr}{Expression to evaluate}

\item{threshold}{Memory threshold in MB}
}
\value{
Result of expression
}
\description{
Executes operation with memory monitoring and cleanup
}
\keyword{internal}
