% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectRank.R
\name{calculate_significant_components}
\alias{calculate_significant_components}
\title{Calculate Significant Components}
\usage{
calculate_significant_components(vectors, k, n.genes, max_p_value)
}
\description{
Calculate Significant Components
}
\keyword{internal}
