% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{csfnmf-class}
\alias{csfnmf-class}
\title{CSFNMF Class}
\description{
Main class for Constrained Supervised Factorization NMF analysis
}
\section{Slots}{

\describe{
\item{\code{matrices}}{Current working matrices (RefDataList)}

\item{\code{annotation}}{Data frame of annotations}

\item{\code{train_object}}{Training results and parameters}

\item{\code{H}}{Final H matrix}

\item{\code{W}}{Final W matrix}
}}

