% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\name{h.baron_dataset}
\alias{h.baron_dataset}
\title{Load Baron Human Pancreas Dataset (Deprecated)}
\usage{
h.baron_dataset()
}
\value{
A list containing the dataset
}
\description{
This function has been renamed. Please use \code{hBaronDataset()} instead.
}
\keyword{internal}
