% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R, R/classes.R
\docType{class}
\name{helpmat-class}
\alias{helpmat-class}
\title{Helper Matrices Class}
\description{
Class for storing intermediate matrices used in NMF computation. These matrices
are used to optimize the factorization process and enforce constraints.

Class for storing intermediate matrices used in NMF computation
}
\section{Slots}{

\describe{
\item{\code{A}}{Matrix for within-class scatter computation}

\item{\code{B}}{Matrix for between-class scatter computation}

\item{\code{P}}{Projection matrix for class relationships}

\item{\code{M}}{Regularization matrix for W updates}

\item{\code{N}}{Normalization matrix for class weights}

\item{\code{BP}}{Combined B and P matrices for efficiency}

\item{\code{BP_posneg}}{List containing positive and negative components of BP matrix}

\item{\code{Hconst}}{List of constants used in H matrix updates}

\item{\code{A}}{First matrix}

\item{\code{B}}{Second matrix}

\item{\code{P}}{P matrix}

\item{\code{M}}{M matrix}

\item{\code{N}}{N matrix}

\item{\code{BP}}{BP matrix}

\item{\code{BP_posneg}}{List of positive/negative BP values}

\item{\code{Hconst}}{List of H constants}
}}

