% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{RefDataList-class}
\alias{RefDataList-class}
\title{RefDataList Class}
\description{
Class for storing paired reference and query expression matrices
}
\section{Slots}{

\describe{
\item{\code{ref}}{Reference matrix (sparse Matrix) containing gene expression from labeled cells}

\item{\code{data}}{Query matrix (sparse Matrix) containing gene expression from unlabeled cells}
}}

