% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{traincsfnmf-class}
\alias{traincsfnmf-class}
\title{TrainCSFNMF Class}
\description{
Class for storing training results and parameters
}
\section{Slots}{

\describe{
\item{\code{count.matrices}}{RefDataList object}

\item{\code{annotation}}{Data frame of annotations}

\item{\code{test.annotation}}{Data frame of test annotations}

\item{\code{rank}}{Rank object}

\item{\code{H}}{H matrix}

\item{\code{W}}{W matrix}

\item{\code{constants}}{Helper matrices}

\item{\code{hyper_para}}{List of hyperparameters}

\item{\code{init_method}}{Character indicating initialization method}

\item{\code{max.iter}}{Numeric maximum iterations}

\item{\code{loss}}{Numeric loss value}

\item{\code{accuracy}}{Numeric accuracy value}
}}

