% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{traincsfnmf-class}
\alias{traincsfnmf-class}
\title{TrainCSFNMF Class}
\description{
Class for storing training results and parameters
}
\section{Slots}{

\describe{
\item{\code{matrices}}{Current working matrices (RefDataList)}

\item{\code{annotation}}{Training data annotations}

\item{\code{test_annotation}}{Test data annotations}

\item{\code{rank}}{Numeric rank value}

\item{\code{H}}{H matrix}

\item{\code{W}}{W matrix}

\item{\code{constants}}{Helper matrices}

\item{\code{parameters}}{List of parameters (hyperparameters, initialization method, etc.)}

\item{\code{results}}{List of results (loss, accuracy)}
}}

