% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{update_traincsfnmf-class}
\alias{update_traincsfnmf-class}
\title{Training Results Class}
\description{
Class for storing training results with updates
}
\section{Slots}{

\describe{
\item{\code{base}}{Base training results}

\item{\code{updates}}{List of update results}
}}

