% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainModelCSFNMF.R
\name{UpdateHyperpara}
\alias{UpdateHyperpara}
\title{Update Hyperparameters}
\usage{
UpdateHyperpara(train_object, singler_pred, pra_to_update)
}
\arguments{
\item{train_object}{Training object}

\item{singler_pred}{SingleR predictions}

\item{pra_to_update}{Parameters to update}
}
\value{
List of updated parameters
}
\description{
Update Hyperparameters
}
\keyword{internal}
